move peer_id from secure_storage_config as a method of NetworkConfig
Closes: #4582
Approved by: davidiw
