+++ config/management/src/smoke_test.rs
@@ -4,8 +4,8 @@
 use crate::{constants, layout::Layout, storage_helper::StorageHelper};
 use config_builder::{BuildSwarm, SwarmConfig};
 use libra_config::config::{
-    peer_id, Identity, NodeConfig, OnDiskStorageConfig, RoleType, SecureBackend,
-    SeedPeersConfig, WaypointConfig, HANDSHAKE_VERSION,
+    Identity, NodeConfig, OnDiskStorageConfig, RoleType, SecureBackend, SeedPeersConfig,
+    WaypointConfig, HANDSHAKE_VERSION,
 };
 use libra_crypto::ed25519::Ed25519PrivateKey;
 use libra_secure_storage::{CryptoStorage, KVStorage, Value};
@@ -211,7 +211,7 @@ fn attach_validator_full_node(validator_config: &mut NodeConfig) -> NodeConfig {
     let v_vfn_pub_key = v_vfn.identity_key().public_key();
     let v_vfn_network_address =
         v_vfn_network_address.append_prod_protos(v_vfn_pub_key, HANDSHAKE_VERSION);
-    let v_vfn_id = peer_id(v_vfn);
+    let v_vfn_id = v_vfn.peer_id();
     let mut seed_peers = SeedPeersConfig::default();
     seed_peers.insert(v_vfn_id, vec![v_vfn_network_address]);
