+++ network/builder/src/builder.rs
@@ -13,7 +13,7 @@ use channel::{self, libra_channel, message_queues::QueueStyle};
 use futures::stream::StreamExt;
 use libra_config::{
     chain_id::ChainId,
-    config::{peer_id, DiscoveryMethod, NetworkConfig, RoleType, HANDSHAKE_VERSION},
+    config::{DiscoveryMethod, NetworkConfig, RoleType, HANDSHAKE_VERSION},
     network_id::{NetworkContext, NetworkId},
 };
 use libra_crypto::x25519;
@@ -531,7 +531,7 @@ pub fn setup_network(
         .expect("Failed to start runtime. Won't be able to start networking.");
 
     let identity_key = config.identity_key();
-    let peer_id = peer_id(config);
+    let peer_id = config.peer_id();
 
     let mut network_builder = NetworkBuilder::new(
         runtime.handle().clone(),
